feat(notification_permission): Add check permission channel and improve organisation of platform channel code in flutter
