+++ yaki_mobile/lib/main.dart
@@ -2,9 +2,9 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:yaki/app.dart';
 import 'package:flutter/services.dart';
+import 'package:yaki/channels.dart';
 import 'package:yaki/presentation/state/dio/dio_interceptor.dart';
 
 void main() async {
@@ -39,19 +39,8 @@ void main() async {
     ),
   );
 
-  // Notification channel
-  const platform = MethodChannel('com.xpeho.yaki/notification');
-
-  platform.setMethodCallHandler((MethodCall call) async {
-    switch (call.method) {
-      case 'areNotificationsActivated':
-        return await areNotificationsActivated();
-      case 'getNotificationParams':
-        return getNotificationParams();
-      default:
-        throw MissingPluginException();
-    }
-  });
+  // Platform channels
+  initChannels();
 }
 
 Future<List<Override>> _overrides() async {
@@ -65,20 +54,4 @@ Future<List<Override>> _overrides() async {
       serverCertificate,
     ),
   ];
-}
-
-Future<bool> areNotificationsActivated() async {
-  SharedPreferences prefs = await SharedPreferences.getInstance();
-  final bool? areNotificationsActivated =
-      prefs.getBool('areNotificationsActivated');
-  return areNotificationsActivated ?? false;
-}
-
-Map<String, dynamic> getNotificationParams() {
-  return {
-    'title': tr("notificationTitle"),
-    'message': tr("notificationMessage"),
-    'hour': 9,
-    'minute': 0,
-  };
 }
